Add files via upload
This is a self-authored ebook project, developed as a way to explore and test my technical and creative skills. It offers a brief overview of my personal and professional journey so far, blending life experiences, academic growth, and lessons learned in a clear and approachable format.

More than just storytelling, this project represents a hands-on initiative to apply my knowledge in development and digital content structuring combining curiosity, self-learning, and the desire to build something unique.